requirements.txt
Django==5.0.6
django-ninja==1.2.1
